==> notebooks/00-EDA.ipynb <==
# Fetch Data Challenge
You will be provided with a dataset of offers and some associated metadata around the retailers and brands that are sponsoring the offer.

You will also be provided with a dataset of some brands that we support on our platform, and the categories that those products belong
to.

# Acceptance Criteria
- If a user searches for a category (ex. diapers) the tool should return a list of offers that are relevant to that
category.
- If a user searches for a brand (ex. Huggies) the tool should return a list of offers that are relevant to that brand.
- If a user searches for a retailer (ex. Target) the tool should return a list of offers that are relevant to that retailer.
- The tool should also return the score that was used to measure the similarity of the text input with each offer

# What we seek below:
- Detailed responses to each problem, with a focus on the production pipeline surrounding the model.
- Identifies several useful techniques to approach eReceipt classification and entity extraction.
- Demonstrate a knowledge of recent innovations in NLP and a willingness to think about the problem in terms of software engineering rather than an academic exercise

## Data Inspection and EDA

# check current path
import os
import pandas as pd
pd.set_option('display.max_colwidth', None)


# # Load the Drive helper and mount
# from google.colab import drive
# # This will prompt for authorization.
# drive.mount('/content/drive/')
# path_gdrive = '/content/drive/MyDrive/Colab Datasets/Fetch'
# os.chdir(path_gdrive)
# print(os.getcwd())

path = "/Users/anthony/Projects/retailer_nlp_challenger/data"
os.chdir(path)

## Categories.csv

import pandas as pd

df_categories = pd.read_csv('data/categories.csv')
df_categories

# how many categories? how many category id's?
df_categories.info()

cat_features = df_categories.select_dtypes(include=['object']).columns.tolist()

for i in cat_features:
    print(i,len(df_categories[i].value_counts()))

# find if there're overlaps between categories and upper categories
all_categories = df_categories['PRODUCT_CATEGORY'].unique().tolist()
upper_categories = df_categories['IS_CHILD_CATEGORY_TO'].unique().tolist()

common_categories = [i for i in all_categories if i in upper_categories]
print(common_categories)
print(len(common_categories))

upper_categories

import re

def search_list(pattern, string_list):
    matches = []
    for string in string_list:
        if re.search(pattern, string, flags=re.IGNORECASE):
            matches.append(string)
    if len(matches) == 0:
        print("no matches")
        return None
    else:
        return matches

search_list(r'Alcohol', all_categories)

df_categories[df_categories["PRODUCT_CATEGORY"].eq('Wine')]

df_categories[df_categories["IS_CHILD_CATEGORY_TO"].eq('Alcohol')]

For categories.csv, we know:
- There are 118 categories in total
- There are only 118-9=109 child categories

## retailer.csv

df_offers = pd.read_csv('data/offer_retailer.csv')
df_offers

df_offers.info()

# notice a column with NaN; check null counts
df_offers.isnull().sum()

# check categorical feature uniqueness
for i in df_offers.columns.tolist():
    print(i,len(df_offers[i].value_counts()))

# find if there're overlaps between retailers and brands
retailers = df_offers['RETAILER'].unique().tolist()
brands = df_offers['BRAND'].unique().tolist()

# what brands are also the retailers
[i for i in retailers if i in brands]

# I call them b_retailer
# does such retailer provide different brand offers?
b_retailer = [i for i in retailers if i in brands]

# group by b_retailer and count the distinct brands they have
df_offers[df_offers['RETAILER'].isin(b_retailer)].groupby('RETAILER')['BRAND'].nunique()

df_offers[df_offers["RETAILER"].eq("SAMS CLUB")] #["BRAND"].value_counts()

retailers2 = [i for i in retailers if isinstance(i, str)]
[i for i in retailers2 if ".com" in i.lower()]

## Brand

df_brand = pd.read_csv('data/brand_category.csv')
df_brand

df_brand.info()

# why is one BRAND null?
df_brand[df_brand["BRAND"].isnull() == True]

Clearly the brands should be cleaned because some brand names do not make sense. For example, you might have names like `3M` or `Forever21`, but no brand names are `21`.

grouped_df = df_brand.groupby('BRAND')['BRAND_BELONGS_TO_CATEGORY'].unique().reset_index()
print(grouped_df.shape)
grouped_df

# clean the rows based on BRAND: (1) NaN (2) only has numbers
df_brand = df_brand.dropna(subset=['BRAND'])
df_brand = df_brand[~df_brand['BRAND'].str.isnumeric()]
df_brand.shape

# check categorical feature uniqueness
for i in df_brand.columns.tolist():
    print(i,len(df_brand[i].value_counts()))

grouped_df = df_brand.groupby('BRAND')['BRAND_BELONGS_TO_CATEGORY'].unique().reset_index()
print(grouped_df.shape)
grouped_df

grouped_df.BRAND_BELONGS_TO_CATEGORY.apply(len).describe()

So, one brand might correspond to multiple categories in a few cases

filtered_df = grouped_df[grouped_df['BRAND_BELONGS_TO_CATEGORY'].apply(len) > 1]
filtered_df

One brand can belong to mutiple categories. And one category surely has mutiple brands.

## Create new table

df_offers = pd.read_csv('data/brand_offer_cleaned.csv')
df_offers.sample(5)

df_offers.shape

# reload data
df_brand = pd.read_csv('data/brand_category.csv')
print(f'Brand original shape: {df_brand.shape}')

# clean the rows based on BRAND: (1) NaN (2) only has numbers
df_brand = df_brand.dropna(subset=['BRAND'])
df_brand = df_brand[~df_brand['BRAND'].str.isnumeric()]
print(f'Brand cleaned shape: {df_brand.shape}')

df_brand.sample(5)

# find for each category, what brands does it have
grouped_df = df_brand.groupby('BRAND_BELONGS_TO_CATEGORY')['BRAND'].unique().reset_index()
print(grouped_df.shape)
grouped_df = grouped_df.rename(columns={'BRAND_BELONGS_TO_CATEGORY': 'CATEGORY'})
grouped_df.head()

# write into a dict
brand_belong_category_dict = {row['CATEGORY']: row['BRAND'].tolist() for index, row in grouped_df.iterrows()}
brand_belong_category_dict

# save the dict to json
import json
with open('data/brand_belong_category_dict.json', 'w') as fp:
    json.dump(brand_belong_category_dict, fp)

import pickle
# get brand names we have existing offers for
offered_brands = df_offers.BRAND.unique().tolist()

# # write them to csv 
# with open('data/offered_brands.pkl', 'wb') as f:
#     pickle.dump(offered_brands, f)
