==> notebooks/03-Colab_spacy_train.ipynb <==
# Load the Drive helper and mount
from google.colab import drive
import os
# This will prompt for authorization.
drive.mount('/content/drive/')
path_gdrive = '/content/drive/MyDrive/Colab Datasets/Fetch'
os.chdir(path_gdrive)
print(os.getcwd())

!ls

!pwd

!pip install spacy-transformers

## Train on brand NER

!python -m spacy train config_transformer.cfg --output models/ --paths.train corpus/train.spacy --paths.dev corpus/dev.spacy --gpu-id 0

## Test

import spacy
import pandas as pd
nlp = spacy.load('models/model-last')
val = pd.read_csv('./data/val.csv')

from spacy.tokens import DocBin
import srsly
import spacy
import spacy
from spacy.training import offsets_to_biluo_tags

pd.set_option('display.max_colwidth', None)

print(val.shape)
val.head()

for i,row in val.iterrows():
    doc = nlp(row['OFFER'])
    print("nReal brand : ", row['BRAND'])
    print("nReal retailer: ", row['RETAILER'])
    spacy.displacy.render(doc, style="ent", jupyter=True) # display in Jupyter
    print('---'*10)

## Push to HuggingFace

!pip install spacy-huggingface-hub

!python -m spacy huggingface-hub --help

!huggingface-cli login

!ls models/

!python -m spacy package models/model-last ./models/output --build wheel --name fetch_ner_spacy_tsf

!python -m spacy huggingface-hub push models/output/en_fetch_ner_spacy_tsf-0.0.0/dist/en_fetch_ner_spacy_tsf-0.0.0-py3-none-any.whl

## Test on Huggingface model

!pip install https://huggingface.co/hjianganthony/en_fetch_ner_spacy_tsf/resolve/main/en_fetch_ner_spacy_tsf-any-py3-none-any.whl

# Using spacy.load().
import spacy
nlp = spacy.load("en_fetch_ner_spacy_tsf")

# Importing as module.
import en_fetch_ner_spacy_tsf
nlp = en_fetch_ner_spacy_tsf.load()

from spacy.tokens import DocBin
import srsly
# from spacy.training import offsets_to_biluo_tags

# Load the Drive helper and mount
from google.colab import drive
import os
# This will prompt for authorization.
drive.mount('/content/drive/')
path_gdrive = '/content/drive/MyDrive/Colab Datasets/Fetch'
os.chdir(path_gdrive)
print(os.getcwd())

import pandas as pd
val = pd.read_csv('./data/val.csv')

pd.set_option('display.max_colwidth', None)

print(val.shape)
val.head()

for i,row in val.iterrows():
    doc = nlp(row['OFFER'])
    print("nReal brand : ", row['BRAND'])
    print("nReal brand : ", row['RETAILER'])
    spacy.displacy.render(doc, style="ent", jupyter=True) # display in Jupyter
    print('---'*10)

text = "Butterball, select varieties, spend 10 at King Soopers"
doc = nlp(text)
spacy.displacy.render(doc, style="ent", jupyter=True) # display in Jupyter
