==> notebooks/04-Search Processing.ipynb <==
import pandas as pd
import numpy as np

from spacy.tokens import DocBin
import srsly
import spacy
from spacy.training import offsets_to_biluo_tags

# Importing as module.
import en_fetch_ner_spacy_tsf
nlp = en_fetch_ner_spacy_tsf.load()

import os
path = "/Users/anthony/Projects/retailer_nlp_challenger/data"
os.chdir(path)

pd.set_option('display.max_colwidth', None)

## Load Operation Data

import pickle, json

path1 = "brand_belong_category_dict.json"
path2 = "product_upper_category_dict.json"
path3 = "offered_brands.pkl"
path4 = "offer_retailer.csv"

with open(path1, 'r') as f:
    brand_belong_category_dict = json.load(f)

with open(path2, 'rb') as f:
    category_dict = json.load(f)

with open(path3, 'rb') as f:
    offered_brands = pickle.load(f)

df_offers_brand_retailer = pd.read_csv(path4)
df_offers_brand_retailer.head()

category_dict

brand_belong_category_dict

Assumptions:
- search input will be simple sentences or keywords (e.g. `beer`, or `beers discount Amazon`)
- users won't serach upper_category because the APP already has UI for them to filter
- `PRODUCT_CATEGORY` are not to spefici brands, so a customer might search `beer` or `white beer`
- the search functions I am using will clean the text to lower case. This is because I don't know if the `PRODUCT_CATEGORY` data gives me the write permission

Issues with dataset:
- `Alchol` should be renamed as drinks. It contains more than alcholic drinks `[Beer, Malt Beverages, Hard Seltzers, Sodas, Waters, Lemonades & Teas, Hard Ciders, Wine, Spirits]`

## Helper functions

import nltk
nltk.download('stopwords')
from nltk.corpus import stopwords
stop_words = set(stopwords.words('english'))

additional_stop_words = {'pack'}
stop_words.update(additional_stop_words)


# helper functions
from typing import List, Dict, Tuple
import re

from sklearn.feature_extraction.text import CountVectorizer
from sklearn.metrics.pairwise import cosine_similarity

def single_text_cleaner(text: str, remove_stopwords: bool=False, upper_case: bool = False, remove_punctuation: bool=True) -> str:
    """Clean one single text input. By default it will convert text to lower case"""
    if upper_case:
        text = text.upper()
    else:
        text = text.lower()
    if remove_punctuation:
        text = re.sub(r'[^a-z\s]', '', text)
    if remove_stopwords:
        words = text.split()
        words = [word for word in words if word not in stop_words]
        text = ' '.join(words)
    return text

def list_text_cleaner(texts: List[str], upper_case: bool = False, remove_stopwords: bool = False, remove_punctuation: bool=True) -> List[str]:
    """Takes in a list of strings and returns a list of cleaned strings without stop words. 
    Current tasks: 
    - remove non-alphabetical characters
    - converting to lower cases
    - remove stop words (optional)"""
    cleaned_texts = [single_text_cleaner(text, remove_stopwords, upper_case, remove_punctuation) for text in texts]
    return cleaned_texts

def match_product_category(s1: list[str], s2: list[str]) -> str:
    """Find if items of a list is in one list of product categories"""
    return next((p for c in s1 for p in s2 if c in p), None) # this will stop after finding first match, which saves time

def find_category(search_input: str, search_dict: Dict) -> str:
    """Find the category of a search input based on a dictionary of categories
    Args:
    - search_input: a string
    - search_dict: a dictionary of product categories
    """
    search_list = list_text_cleaner(re.split(r'[,\s]+', search_input), remove_stopwords=True)
    search_list = [c for c in search_list if len(c)>0] # sometimes there are empty strings
    matched_category = False
    for k, v in search_dict.items():
        v = list_text_cleaner(v, remove_punctuation=False)
        search_results = match_product_category(search_list, v)
        if search_results is not None:
            matched_category = True
            return k, search_results
        else:
            # print(f'Function find_category: No category {k} has matched for input: {search_input}') 
            continue
    if not matched_category:
        print(f'Function find_category: No category has matched for input: {search_input}')
        return None
    

def check_entity(search_input) -> bool:
    """Takes in a search input and checks if it contains any entities"""
    doc = nlp(search_input)
    if len(doc.ents) > 0:
        return doc
    else:
        return False

def get_cosine_sim(input_text: str, texts: List[str]) -> pd.DataFrame:
    """Calculate the cosine similarity of the input text against a list of texts
    Takes in:
    - input_text: a string
    - texts: a list of strings
    Returns a dataframe with two columns: Sentence Text and Cosine Similarity Score
    """
    input_text_cleaned = list_text_cleaner([input_text], remove_stopwords=True)[0]
    cleaned_texts = list_text_cleaner(texts, remove_stopwords=True)
    all_texts = [input_text_cleaned] + cleaned_texts
    vectors = get_vectors(*all_texts)
    sim_matrix = cosine_similarity(vectors)
    # Get the similarity scores of the input_text against all other texts
    sim_scores = sim_matrix[0, 1:]
    data = {'OFFER': texts, 'Cosine Similarity Score': sim_scores}
    df = pd.DataFrame(data)
    df = df.sort_values(by='Cosine Similarity Score', ascending=False).reset_index(drop=True)
    return df

def get_vectors(*strs: str) -> np.ndarray:
    text = list(strs)
    vectorizer = CountVectorizer()
    vectorizer.fit(text)
    return vectorizer.transform(text).toarray()

def jaccard_similarity(s1: List[str], s2: List[str]) -> float:
    """Takes in two lists and returns the Jaccard similarity score (3 digits)"""
    intersection = set(s1).intersection(set(s2))
    n = len(intersection)
    score = round(n / (len(s1) + len(s2) - n), 3)
    return score

def get_jaccard_sim(input_text: str, texts: List[str]) -> pd.DataFrame:
    """Calculate the Jaccard similarity of the input text against a list of texts
    Takes in:
    - input_text: a string
    - texts: a list of strings
    Returns a dataframe with two columns: Sentence Text and Jaccard Similarity Score
    """
    cleaned_input_text = list_text_cleaner([input_text], remove_stopwords=True)[0].split()
    cleaned_texts = list_text_cleaner(texts, remove_stopwords=True)
    
    jaccard_scores = [jaccard_similarity(cleaned_input_text, text.split()) for text in cleaned_texts]
    
    data = {'OFFER': texts, 'Jaccard Similarity Score': jaccard_scores}
    df = pd.DataFrame(data)
    # sort based on the similarity score
    df = df.sort_values(by='Jaccard Similarity Score', ascending=False).reset_index(drop=True)
    return df

def find_column(df: pd.DataFrame, keyword: str) -> str:
    """Function to find the first column containing a specific keyword. Note that we assume there will only be one score at most for a similarity score dataframe"""
    cols = [col for col in df.columns if keyword.lower() in col.lower()]
    return cols[0] if cols else None

def extract_similar_offers(data: pd.DataFrame, threshold: float = 0.0) -> pd.DataFrame:
    """Takes in the results from get_cosine_sim() and get_jaccard_sim(); returns a dataframe of similar offers with scores > threshold"""
    score = find_column(data, 'score')
    similar_offers = data[data[score] >= threshold]
    return similar_offers

def category_to_brand(category: str, offered_brands: List, brand_belong_category_dict: Dict) -> List[str]:
    """Use case: when a user searches for a category, we return a list of brands in that category"""
    # checks if the category is in the dictionary keys
    if category.upper() in brand_belong_category_dict.keys():
        search_brands = brand_belong_category_dict[category.upper()] # becase all keys are in upper case
        result = list(set(search_brands) & set(offered_brands))
        print(f"Function category_to_brand | Found {category} in offered brand") 
        return result 
    else:
        print(f"Function category_to_brand | No offered brand is found in {category}")
        return None

class CatchErros(Exception):
    class ParamsInputError(Exception):
        pass
    class SearchFailedError(Exception):
        pass
    class UnknownError(Exception):
        pass


def offer_finder_by_category(search_input: str, search_category_tuple: Tuple, category_dict: Dict, offered_brands: List, 
                             brand_belong_category_dict: Dict, score: str, threshold: float = 0.0) -> pd.DataFrame:
    """Find offers based on a category identified from search input.
    Args:
    - search_input: a string
    - search_category_tuple: a tuple of (upper_category, product_category)
    - category_dict: a dictionary of categories. Keys are upper categories and values are lists of product categories
    - offered_brands:  a dataframe of offers (OFFER, BRAND, RETAILER) that are avaialble in our database
    - brand_belong_category_dict: a dictionary of brands and the categories they belong to
    - score: a string of either 'cosine' or 'jaccard'
    - threshold: a float between 0 and 1

    Returns a dataframe of similar offers, ordered by highest score
    """
    # we assume people just search one category at a time
    # search_category_tuple = find_category(search_input, category_dict)
    product_category, upper_category = search_category_tuple[1], search_category_tuple[0] # ('Alcohol', 'beer')
    print(f'Function offer_finder_by_category | Found items:\n- Search input: {search_input}\n- Product category: {product_category}\n- Upper category: {upper_category}')
    potential_brands = category_to_brand(product_category, offered_brands, brand_belong_category_dict)
    if potential_brands is not None:
        potential_offers = offers[offers['BRAND'].isin(potential_brands)]['OFFER'].tolist()
        if score == 'cosine':
            cos_sim_score = get_cosine_sim(search_input, potential_offers)
            output = extract_similar_offers(cos_sim_score, threshold)
        elif score == 'jaccard':
            jaccard_sim_score = get_jaccard_sim(search_input, potential_offers)
            output = extract_similar_offers(jaccard_sim_score, threshold)
        elif score not in ['cosine', 'jaccard']:
            raise ValueError(f'Please enter a valid score: cosine or jaccard; Not {score}')
        else: # this means something else is worng
            raise UnknownError(f'Something must be broken. Please try again.')
        return output
    else:
        potential_product_categories = category_dict[upper_category]
        msg = f'{product_category} is not found. Do you wanna take a look at these similar offers in {upper_category}?\n We have: {potential_product_categories}' # we can still calculate similarity but this is computationally expensive
        print(msg)
        return None

def offer_finder_by_entity(search_input: str, entities: Tuple, offers_data: pd.DataFrame, score: str, threshold: float=0.0) -> pd.DataFrame:
    """Find offers based on entities identified from search input.
    Args:
    - search_input: a string
    - entities: a tuple of entities
    - offers_data: a dataframe of offers (OFFER, BRAND, RETAILER) that are avaialble in our database
    - score: a string of either 'cosine' or 'jaccard'
    - threshold: a float between 0 and 1

    Returns a dataframe of similar offers, ordered by highest score
    """
    collects = [] # collect all the results if there are more than one entity
    for ent in entities:
        ent_name, ent_label = ent.text, ent.label_
        print(f'Function offer_finder_by_entity | Found entity: {ent_name} with label: {ent_label}')
        # filter offers by entity
        df_tmp = offers_data[offers_data[ent_label.upper()] == ent_name.upper()]
        if df_tmp.shape[0] > 0:
            print(f'Function offer_finder_by_entity | Found {df_tmp.shape[0]} offer(s) for the brand/retailer: {ent_name}')
            potential_offers = df_tmp['OFFER'].drop_duplicates().tolist()
            if score == 'cosine':
                cos_sim_score = get_cosine_sim(search_input, potential_offers)
                output = extract_similar_offers(cos_sim_score, threshold)
            elif score == 'jaccard':
                jaccard_sim_score = get_jaccard_sim(search_input, potential_offers)
                output = extract_similar_offers(jaccard_sim_score, threshold)
            elif score not in ['cosine', 'jaccard']:
                raise ValueError(f'Please enter a valid score: cosine or jaccard; Not {score}')
            else: # this means something else is worng
                raise UnknownError(f'Something must be broken. Please try again.')
            collects.append(output)
        else:
            print(f'Function offer_finder_by_entity | No offer is found for the brand/retailer: {ent_name}')

    if len(collects) > 0:
        final_output = pd.concat(collects, ignore_index=True)# they should be using the same similarity score
        score = find_column(collects[0], 'score') 
        final_output = final_output.sort_values(by=score, ascending=False).reset_index(drop=True) # sort final_output by score
        return final_output
    elif len(collects) == 1:
        return collects[0]
    else:
        print('###'*5 + 'FINAL SEARCH RESULTS' + '###'*5)
        print('Function offer_finder_by_entity | No offer is found for any of the entities.')
        return None


def main(search_input: str, offers: pd.DataFrame, category_dict: Dict, brand_belong_category_dict: Dict, score: str, score_threshold: float = 0.0):
    """Main function. Takes in a serach_input and decide whether it can find entities or not. Then excecute the appropriate functions
    Inputs:
    - search_input: a string that a user enters
    - offers: a dataframe of offers (OFFER, BRAND, RETAILER) that are avaialble in our database
    - category_dict: a dictionary of categories. Keys are upper categories and values are lists of product categories
    - brand_belong_category_dict: a dictionary of brands and the categories they belong to
    - score: a string of either 'cosine' or 'jaccard'
    - score_threshold: a float between 0 and 1

    Returns a dataframe of similar offers, ordered by highest score
    """
    print(f'Function main | Search input: {search_input}')
    check_ent = check_entity(search_input)
    if not check_entity(search_input): # no entities found
       # check category
       cat_check = find_category(search_input, category_dict)
       if cat_check is None:
           raise SearchFailedError('No brand/retailer/category is found. Please try again.')
       else:
            # we assume people just search one category at a time
            cat_tuple = cat_check # ('Alcohol', 'beer')
            search_results = offer_finder_by_category(search_input, cat_tuple, category_dict, offers, brand_belong_category_dict, score, score_threshold)
            return search_results
    else:
        entities = check_ent.ents # entities will be a tuple anyways
        print(f'Found {len(entities)} entity object(s) in the search input.')
        search_results = offer_finder_by_entity(search_input, entities, offers, score, score_threshold)
        if search_results is None:
            print('No offers matched retailer/category is found. Now trying to recommend based on category.')
            cat_check = find_category(search_input, category_dict)
            if cat_check is None:
                raise SearchFailedError('No brand/retailer/category is found. Please try again.')
            else:
                cat_tuple = cat_check
                search_results = offer_finder_by_category(search_input, cat_tuple, category_dict, offers, brand_belong_category_dict, score, score_threshold)
        return search_results
            

## Playground to test

### Main Function Test

test = "jaja"

main(test, df_offers_brand_retailer, category_dict, brand_belong_category_dict, 'cosine', 0.0)

# test = "Simply Spiked Lemonade 12 pack at Walmart"
test = "EGGLIFE 12 pack"

main(test, df_offers_brand_retailer, category_dict, brand_belong_category_dict, 'cosine', 0.0)

test = "Simply Spiked Lemonade 12 pack at Walmart"
# test = "EGGLIFE 12 pack"

main(test, df_offers_brand_retailer, category_dict, brand_belong_category_dict, 'cosine', 0.3)

# test = "Sara Lee bread, select varieties, buy 2 at Walmart"
test = "Back to the Roots Garden Soil, 1 cubic foot, at Lowe's Home Improvement"

main(test, df_offers_brand_retailer, category_dict, brand_belong_category_dict, 'cosine', 0.3)

test = "Back to the Roots Garden Soil, 1 cubic foot, at Lowe's Home Improvement"

main(test, df_offers_brand_retailer, category_dict, brand_belong_category_dict, 'jaccard', 0.1)

test = "Membership at Costco"

main(test, df_offers_brand_retailer, category_dict, brand_belong_category_dict, 'cosine', 0.2)

main(test, df_offers_brand_retailer, category_dict, brand_belong_category_dict, 'jaccard', 0.1)

### Functions test

# test case: find category

category_dict = brand_belong_category_dict

test1 = "White German Beer"
test2 = ["Beer", "Malt Beverages", "Hard Seltzers", "Sodas", "Waters", "Lemonades & Teas", "Hard Ciders", "Wine", "Spirits"]

find_category(test1, category_dict)


test_text = "Klondike Cones at Walmart"
# test_text = "Klondike Cones"
# test_test = "M&M's candy discount at Target"


doc = nlp(test_text)
spacy.displacy.render(doc, style="ent", jupyter=True) # display in Jupyter

# [(ent.text, ent.label_) for ent in text.ents]

# len(check_entity(test_text).ents)

# type(check_entity(test_text).ents)

for ent in doc.ents:
    print(f'Entity: {ent.text} | Label: {ent.label_}')

# test on Cosine Similarity for search_offer comparison

input_text = "Membership at Costco"
texts = ["Sara Lee bread, select varieties, buy 2 at Walmart", 
         "Back to the Roots Garden Soil, 1 cubic foot, at Lowe's Home Improvement", 
         "When you join Costco as a Gold Star Member", 
         "this is the fourth text"]

cosine_sim_scores = get_cosine_sim(input_text, texts)
cosine_sim_scores

# test on Jaccard Similarity for search_offer comparison

input_text = "Membership at Costco"
texts = ["Sara Lee bread, select varieties, buy 2 at Walmart", 
         "Back to the Roots Garden Soil, 1 cubic foot, at Lowe's Home Improvement", 
         "When you join Costco as a Gold Star Member", 
         "this is the fourth text"]

result = get_jaccard_sim(input_text, texts)
result

# test case on extract_similar_offers
extract_similar_offers(get_jaccard_sim(input_text, texts))

extract_similar_offers(get_cosine_sim(input_text, texts))

# df_offers = pd.read_csv(path + 'brand_offer_cleaned.csv')
# df_offers.sample(10)

### Similarity match test

Operation data test

print(f'offered_brands: {type(offered_brands)}')
print(f'brand_belong_category_dict: {type(brand_belong_category_dict)}')

offered_brands[:5]

brand_belong_category_dict['Baby Bathing'.upper()][:5]

# test case: I have a category, now I want to find all the brands (with offers) that belong to this category
# so I need to find a intersection of two lists

category = 'Packaged Meals & Sides'
search_brands = brand_belong_category_dict[category.upper()]

# find the intersection of two lists
result = list(set(search_brands) & set(offered_brands))
result

## Build Category Dictionary

As checks (mannually and EDA), there're no cases for one product_category that belongs to more than one categories. So, we can build a simple category dictionary.

df_categories = pd.read_csv('categories.csv')
print(df_categories.shape)

# rename "IS_CHILD_CATEGORY_TO" to "UPPER_CATEGORY"
df_categories.rename(columns={'IS_CHILD_CATEGORY_TO': 'UPPER_CATEGORY'}, inplace=True)
display(df_categories.head())

# group by UPPER_CATEGORY and get DISTINCT PRODUCT_CATEGORY
grouped_df = df_categories.groupby('UPPER_CATEGORY')['PRODUCT_CATEGORY'].unique().reset_index()
print(grouped_df.shape)
display(grouped_df.head())

# generate a dictionary of UPPER_CATEGORY (key) and PRODUCT_CATEGORY (value)'
category_dict = {row['UPPER_CATEGORY']: row['PRODUCT_CATEGORY'].tolist() for index, row in grouped_df.iterrows()}

# change all keys in category_dict to upper case
category_dict = {k.upper(): v for k, v in category_dict.items()}

# clear all non-str values for all keys, values in category_dict
print(f'Original length of category_dict: {len(category_dict)}')
category_dict_cleaned = {k: [i for i in v if isinstance(i, str)] for k, v in category_dict.items()}
print(f'Cleaned length of category_dict: {len(category_dict_cleaned)}')

# write to json
# import json
# with open('product_upper_category_dict.json', 'w') as fp:
#     json.dump(category_dict_cleaned, fp)

df_brand_to_category = pd.read_csv('brand_category.csv')
df2 = df_brand_to_category
print(df2.shape)
# display(df2.head())

# rename "IS_CHILD_CATEGORY_TO" to "UPPER_CATEGORY"
df2.rename(columns={'BRAND_BELONGS_TO_CATEGORY': 'PRODUCT_CATEGORY'}, inplace=True)
display(df2.head())

# group by UPPER_CATEGORY and get DISTINCT PRODUCT_CATEGORY
grouped_df = df2.groupby('PRODUCT_CATEGORY')['BRAND'].unique().reset_index()
print(grouped_df.shape)
display(grouped_df.head())

# generate a dictionary of UPPER_CATEGORY (key) and PRODUCT_CATEGORY (value)'
brand_belong_category_dict = {row['PRODUCT_CATEGORY']: row['BRAND'].tolist() for index, row in grouped_df.iterrows()}

# change all keys in category_dict to upper case
brand_belong_category_dict = {k.upper(): v for k, v in category_dict.items()}

# clear all non-str values for all keys, values in category_dict
print(f'Original length of brand_belong_category_dict: {len(brand_belong_category_dict)}')
brand_belong_category_dict_cleaned = {k: [i for i in v if isinstance(i, str)] for k, v in brand_belong_category_dict.items()}
print(f'Cleaned length of category_dict: {len(brand_belong_category_dict_cleaned)}')

# # write to json
# import json
# with open('brand_belong_category_dict.json', 'w') as fp:
#     json.dump(brand_belong_category_dict_cleaned, fp)
